requirement.txt
Flask
razorpay
python-dotenv
nltk
fuzzywuzzy[speedup]
